requirements.txt
requests>=2.28.0
beautifulsoup4>=4.12.0
langchain-openai>=0.2.0
python-dotenv>=1.0.0
